Automatically generate preference dialog all properties listed in de.mfo.surfer.util.Preferences
A tab is created for each static inner class of de.mfo.surfer.util.Preferences. For each property of this inner class, a new GridPane row is added to the tab containing the property label and an editor control for the property. Currently supported property types are BooleanProperty and DoubleProperty.