Fix blurry display of gallery info pages
The info pages appeared blurry due to the globally applied scale transformation. This fix is an ugly workaround: The renderings are put into an overlay that is not affected by the global scale transformation. The bounds of the nodes in this overlay are bound to the computed bounding boxes of the gallery info pages with respect to the scene root. This may lead to all kinds of strange behaviour in the future and should only serve as a temporary solution.