Fix blurry display of surface rendering
The RenderArea is placed on top of the scaled scene and its bounds are adjusted according to the scaled placeholder underneath, but with integral position and size. Therefore, the surface rendering will be displayed without any additional scaling if it is the exact same size as the RenderArea, which is the case for all full resolution renderings. To properly scale up the preview images, the display now utilized an ImageView (instead of a Canvas) which has this functionality built in.